Rework chunk generation code for flexibility
This patch changes how `World::Generate` works. The function previously operated on the world coordinates of the chunk to be generated, whereas now it operates on the chunk-scale coordinates of the chunk and takes into account a `blockSize` variable.

`World::GenerateBlock` is updated accordingly to roughly preserve the current world scale and tunnel system periodicity.

Additionally, `Generate` and `GenerateBlock` now each have a single `Point`  argument as their input.