Remove unused glm::pow function call
This patch removes the glm::pow call that used to be relevant in the periodic tunnel system density function. Despite the exponent being 1.0, the function still performs calculations.

The Visual Studio 2013 CPU Usage diagnostic tool was used and it was discovered that the function `libm_sse2_pow_precise` is accountable for 8.8% of the program's CPU usage at startup time.