Eliminate unnecessary operations
This patch performs an early check in the World view distance chunk loop and
skips the current iteration if none of x, y, or z have an absolute value equal
to the current value of d. This has the effect of early-skipping any iterations
that are not part of the ring belonging to the current distance.

In terms of sequence complexity

num operations = 1 squared + 3 squared + 5 squared + ... + n squared

has been reduced to

num operations = n squared + negligible early check operations