Fix in-chunk calculation for negative chunks
This patch applies a floor operation to the result of the division of the
player camera position by the chunk size to correctly calculate the chunk the
camera is currently in for negatively positioned regions. A floor operation
rounds down to the predecessing whole number whereas simply constructing an
integer vector outright truncates the decimal portion of the input vector.