Expose component names to API
This patch adds a required `name` member function to the component base
class. The function is abstract so components are required to implement
it.

A virtual destructor is also added to the component base class to
ensure that all component destructors are called correctly.

Finally, component expressions are added to kludge, as well as reduction
support for expressions like `component.position` and `component.color`.
Finally, reduction support is added for components